+++ sw-sysemu/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [SW-19833] Log PC on coherency hazard
 ### Changed
 - Updated pointers to match release SDK 1.4.0
 - (CI) by default use conan-linux-ubuntu18.04-gcc7 image